Corrected Test

+++ trace-to-events/src/parameters.rs
@@ -51,7 +51,7 @@ pub(crate) struct DifferentialThresholdDiscriminatorParameters {
     /// If set, the derivative is approximated using the central finite difference with
     /// a stencil of 2*r + 1 grid points, where r is `central_fin_diff_radius`,
     /// otherwise the forward finite difference with 2 grid points is used.
-    pub(crate) central_fin_diff_radius: Option<i32>,
+    pub(crate) central_fin_diff_radius: Option<usize>,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
